api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
click==8.1.7
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
environs==10.3.0
exceptiongroup==1.2.0
fastapi==0.109.2
fastapi_cors==0.0.6
flupy==1.2.0
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
marshmallow==3.20.2
numpy==1.26.4
openai==1.11.1
packaging==23.2
pgvector==0.1.8
postgrest==0.15.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
realtime==1.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
vecs==0.4.2
websockets==11.0.3
wrapt==1.16.0
